% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostpower.R
\name{tost_power}
\alias{tost_power}
\title{TOST power and sample size calculations.}
\usage{
tost_power(n = 10, sigma = 1, true_diff = 0, eqv_interval = c(-1, 1),
  alpha = 0.05, paired = FALSE, df = NULL)
}
\arguments{
\item{n}{The number in each group for \code{!paired} (the default);
the number of pairs for \code{paired}}

\item{sigma}{The standard devation: total for \code{!paired}, within for \code{paired}.}

\item{true_diff}{The true (unknown) difference of population means.}

\item{eqv_interval}{The interval within which \code{true_diff} is _equivalant_
considered equivalent (as a proportion of the reference mean).}

\item{alpha}{The significance level of the test.}

\item{paired}{Is this a paired comparison?}

\item{df}{Degress of freedom, (usually not set independent of \code{n}, but can be, see details).}
}
\value{
Power for both equivalence and non-inferiority alternatives
}
\description{
Calculate the power to detect equivalence or non-inferiority for either
paired or unpaired experiments.
}
\examples{
# Reproduce of n=9 part of power curve in Phillips, Fig. 1 on p. 139

delta_mu <- 0:20

sapply(delta_mu, function(x) {
  tost_power(true_diff = x,
    n = 9, sigma = 10,
    eqv_interval = c(-20, 20), df = 7)[['eq']]
  }
)

}
\author{
Dennis L. Malandro, \email{dennismalandro@gmail.com}
}
\references{
Kem F. Phillips (1990) Power of the Two One-Sided Tests Procedure in
Bioequivalence, \emph{J. Pharmacokinetics and Biopharmaceutics} Vol 18, No. 2

Donald J. Schuirmann (1987) A Comparison of the Two One-Sided Tests Procedure and the
Power Approach for Assessing the Equivalence of Average Bioavailability,
 \emph{J. Pharmacokinetics and Biopharmaceutics} Vol 15, No. 6
}
\seealso{
\code{\link{power.t.test}}
}
\keyword{hypothesis-test}
\keyword{power}
\keyword{sample-size}

