% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tostpower.R
\name{tost_power}
\alias{tost_power}
\alias{tost_ss}
\title{TOST power and sample size calculations.}
\usage{
tost_power(n, true_dif = 0, threshold = 0.2, sd = 0.2, sig_level = 0.05,
  paired = FALSE, type = c("equivalence", "non_inferiority"))

tost_ss(power = 0.8, true_dif = 0, threshold = 0.2, sd = 0.2,
  sig_level = 0.05, paired = FALSE, type = c("equivalence",
  "non_inferiority"))
}
\arguments{
\item{n}{The sample size: for \code{paired == TRUE}, the number differences;
for \code{paired == FALSE} (the default), the number in each group.}

\item{true_dif}{The true (ie, unknown) difference in population mean (as a
proportion of the reference mean).}

\item{threshold}{The maximum difference between the two means that can be
considered equivalent (as a proportion of the reference mean).}

\item{sd}{The standard devation (as a proportion of the reference mean).}

\item{sig_level}{The significance (i.e. alpha) level of the test.}

\item{paired}{Is this for paired comparisons?}

\item{type}{Is this designed to show equivalence or simply non-inferiority?}

\item{power}{The power to detect equivalence (or non-inferiority).}
}
\value{
Power to detect equivalence
}
\description{
Calculate the power to detect equivalence or non-inferiority for either
paired or unpaired experiments.
}
\examples{
tost_power(9)

tost_ss(paired = TRUE, type = 'non_inferiority')
}
\author{
Dennis L. Malandro, \email{dennismalandro@gmail.com}
}
\references{
Kem F. Phillips (1990) Power of the Two One-Sided Tests Procedure in
Bioequivalence, \emph{J. Pharmacokinetics and Biopharmaceutics} Vol 18, No. 2

Donald J. Schuirmann (1987) A Comparison of the Two One-Sided Tests Procedure and the
Power Approach for Assessing the Equivalence of Average Bioavailability,
 \emph{J. Pharmacokinetics and Biopharmaceutics} Vol 15, No. 6
}
\seealso{
\code{\link{power.t.test}}
}
\keyword{hypothesis-test}
\keyword{power}
\keyword{sample-size}

